Fabricated cognitive, personality, and vocational interest inventory

The data come from a fabricated cognitive, personality, and vocational interested inventory. This data set can be used to demonstrate regression and structural equation modeling.

interest

Format

A data frame with 250 rows and 33 variables:

gender

1 is female and 2 is male

educ

Years of education

age

Age, in years

vocab

Vocabulary test

reading

Reading comprehension

sentcomp

Sentence completion

mathmtcs

Mathematics

geometry

Geometry

analyrea

Analytical reasoning

socdom

Social dominance

sociabty

Sociability

stress

Stress reaction

worry

Worry scale

impulsve

Impulsivity

thrillsk

Thrill-seeking

carpentr

Carpentry

forestr

Forest ranger

morticin

Mortician

policemn

Police

fireman

Fireman

salesrep

Sales representative

teacher

Teacher

busexec

Business executive

stockbrk

Stock broker

artist

Artist

socworkr

Social worker

truckdvr

Truck driver

doctor

Doctor

clergymn

Clergyman

lawyer

Lawyer

actor

Actor

archtct

Architect

landscpr

Landscaper

Source

http://psych.colorado.edu/~carey/Courses/PSYC7291/ClassDataSets.htm