Profile Analysis for One Sample with Hotelling's T-Square

The paos function implements profile analysis for one sample using Hotelling's T-square.

paos(data, scale = TRUE)

Arguments

data

A data matrix or data frame; rows represent individuals, columns represent variables.

scale

If TRUE (default), variables are standardized by dividing their standard deviations.

Value

A summary table is returned, listing the following two hypothesis:

Hypothesis 1 - Ratios of the means of the variables over the hypothesized mean are equal to 1.

Hypothesis 2 - All of the ratios are equal to each other.

Details

The paos function runs profile analysis for one sample based on the Hotelling's T-square test and tests the two htypothesis. First, the null hypothesis that all the ratios of the variables in the data are equal to 1. After rejecting the first hypothesis, a secondary null hypothesis that all of the ratios of the variables in the data equal to one another (not necessarily equal to 1) is tested.

See also

Examples

if (FALSE) {
data(nutrient) 
paos(nutrient, scale=TRUE)
}