Love and Marriage Survey for Spouses

The spouse data come from a study of love and marriage. A sample of 30 husbands and their wives were asked to respond to the following questions:

Question 1: What is the level of passionate love you feel for your partner?

Question 2: What is the level of passionate love that your partner feels for you?

Question 3: What is the level of companionate love that you feel for your partner?

Question 4: What is the level of companionate love that your partner feels for you?

The responses to all four questions are on a five-point Likert scale where 1 indicates "none at all" and 5 indicates "tremendous amount".

Format

item1

Question 1 with a score ranging from 1 to 5.

item2

Question 2 with a score ranging from 1 to 5.

item3

Question 3 with a score ranging from 1 to 5.

item4

Question 4 with a score ranging from 1 to 5.

spouse

Spouse type. It is either "Husband" or "Wife"

Examples

if (FALSE) {
data(spouse)
}