Criterion-Related Profile Analysis

Implements the criterion-related profile analysis described in Davison & Davenport (2002).

cpa(
  formula,
  data,
  k = 100,
  na.action = "na.fail",
  family = "gaussian",
  weights = NULL
)

Arguments

formula
An object of class formula of the form response ~ terms.

data
An optional data frame, list or environment containing the variables in the model.

k
Corresponds to the scalar constant and must be greater than 0. Defaults to 100.

na.action
How should missing data be handled? Function defaults to failing if missing data are present.

family
A description of the error distribution and link function to be used in the model. See family.

weights
An option vector of weights to be used in the fitting process.

Value

An object of class critpat is returned, listing the following components:

lvl.comp - the level component

pat.comp - the pattern component

b - the unstandardized regression weights

bstar - the mean centered regression weights

xc - the scalar constant times bstar

k - the scale constant

Covpc - the pattern effect

Ypred - the predicted values

r2 - the proportion of variability attributed to the different components

F.table - the associated F-statistic table

F.statistic - the F-statistics

df - the df used in the test

pvalue - the p-values for the test

Details

The cpa function requires two arguments: criterion and predictors. The function returns the criterion-related profile analysis described in Davison & Davenport (2002). Missing data are presently handled by specifying na.action = "na.omit", which performs listwise deletion and na.action = "na.fail", the default, which causes the function to fail. The following S3 generic functions are available: summary(),anova(), print(), and plot(). These functions provide a summary of the analysis (namely, R2 and the level a nd pattern components); perform ANOVA of the R2 for the pattern, the level, and the overall model; provide output similar to lm(), and plots the pattern effect.

References

Davison, M., & Davenport, E. (2002). Identifying criterion-related patterns of predictor scores using multiple regression. Psychological Methods, 7(4), 468-484. DOI: 10.1037/1082-989X.7.4.468.

See also

Examples

if (FALSE) {
data(IPMMc)
mod <- cpa(R ~ A + H + S + B, data = IPMMc)
print(mod)
summary(mod)
plot(mod)
anova(mod)
}